requirements.txt
djangorestframework==3.15.1
Django==3.2.15
